requirements.txt
asgiref==3.4.1
Django==4.1.6
pytz==2021.3
sqlparse==0.4.2
djangorestframework==3.12.4
